refactor: modularize and adopt chatgpt tools
