+++ gategpt/GateGPT/main.js
@@ -3,28 +3,21 @@
  * Please see the LICENSE file.
 *********************************************************************/
 
-const { Client, LocalAuth, MessageMedia, Location } = require('whatsapp-web.js');
+const { Client, LocalAuth, Location } = require('whatsapp-web.js');
 const qrcode = require('qrcode');
 const qrcodeTerminal = require('qrcode-terminal');
 const axios = require('axios');
-const FormData = require('form-data');
 const fs = require('fs');
 const path = require('path');
-const { OpenAI } = require('openai');
 const express = require('express');
 
-/* ──────────────────────────── ⏱  TIMESTAMP PATCH  ─────────────────────────── */
-(() => {
-  const pad = n => String(n).padStart(2, '0');
-  const ts  = () => {
-    const d = new Date();
-    return `${d.getFullYear()}-${pad(d.getMonth() + 1)}-${pad(d.getDate())} ${pad(d.getHours())}:${pad(d.getMinutes())}:${pad(d.getSeconds())}`;
-  };
-  ['log', 'info', 'warn', 'error', 'debug'].forEach(method => {
-    const original = console[method].bind(console);
-    console[method] = (...args) => original(`[${ts()}]`, ...args);
-  });
-})();
+const { initLogging } = require('./logging');
+const { getConfig, reloadConfig, CONFIG_PATH } = require('./config');
+const { sendPushoverNotification } = require('./notifications');
+const { transcribeWithWhisper } = require('./transcription');
+const { askChatGPT } = require('./chatgpt');
+
+initLogging();
 
 /* ─────────────── 🏠  Home-Assistant options → ENV  ─────────────── */
 try {
@@ -34,35 +27,18 @@ try {
 
     Object.entries(opts).forEach(([k, v]) => {
       const key = String(k).toUpperCase();
-
-      // don’t clobber a value that’s already in the real env
       if (process.env[key] === undefined) {
-        process.env[key] =
-          typeof v === 'object' ? JSON.stringify(v) : String(v);
+        process.env[key] = typeof v === 'object' ? JSON.stringify(v) : String(v);
       }
     });
 
-    console.log(
-      `🔧  Loaded ${Object.keys(opts).length} HA option(s) into env vars`
-    );
+    console.log(`🔧  Loaded ${Object.keys(opts).length} HA option(s) into env vars`);
   }
 } catch (err) {
   console.warn('⚠️  Failed to read /data/options.json:', err.message);
 }
 
 /* ──────────────────────── 🔧  CONFIG FILE HANDLING  ───────────────────────── */
-const CONFIG_PATH = path.resolve(__dirname, 'config.json');
-const SAMPLE_CONFIG_PATH = path.resolve(__dirname, 'config.sample.json');
-if (!fs.existsSync(CONFIG_PATH)) {
-  try {
-    fs.copyFileSync(SAMPLE_CONFIG_PATH, CONFIG_PATH);
-    console.warn(`⚠️  config.json missing – copied default settings from config.sample.json`);
-  } catch (err) {
-    console.error('❌ Failed to create config.json from sample:', err.message);
-  }
-}
-let CONFIG = require(CONFIG_PATH);
-
 const DATA_DIR = getConfig('SESSION_DIR', __dirname);
 fs.mkdirSync(DATA_DIR, { recursive: true });
 const QR_PNG_PATH = path.join(DATA_DIR, 'qr.png');
@@ -70,36 +46,16 @@ const SESSION_DIR = path.join(getConfig('SESSION_DIR', __dirname), 'whatsapp-aut
 fs.mkdirSync(SESSION_DIR, { recursive: true });
 
 /* ──────────────────────────── Globals ─────────────────────────── */
-// IDs of messages generated by the bot itself
 const autoMsgIds = new Set();
 
-/**
- * Send a WhatsApp message and remember it as “bot-generated”.
- * Works for text, media, Location, etc.
- */
 async function sendAuto(chat, content, options = {}) {
   const msg = await chat.sendMessage(content, options);
   autoMsgIds.add(msg?.id?._serialized);
-  // Evict after an hour to avoid unbounded growth
   setTimeout(() => autoMsgIds.delete(msg?.id?._serialized), 60 * 60 * 1000);
   return msg;
 }
 
-function getConfig(key, defaultValue = undefined) {
-  const fromEnv = process.env[key];
-
-  if (fromEnv !== undefined) {
-    if (key === 'TRIGGER_KEYWORDS') {
-      return fromEnv.split(/[,]+/).map(s => s.trim()).filter(Boolean);
-    }
-
-    return fromEnv;
-  }
-
-  return CONFIG[key] ?? defaultValue;
-}
-
-// Tiny webserver for QR code
+/* ──────────────────────── Tiny webserver for QR code ─────────────────────── */
 const app = express();
 app.get('/qr.png', (req, res) => res.sendFile(QR_PNG_PATH));
 app.get('/', (req, res) =>
@@ -110,11 +66,10 @@ app.get('/', (req, res) =>
   </body></html>`));
 app.listen(3000);
 
-// Watch for config file changes
-fs.watchFile(CONFIG_PATH, { interval: 1000 }, (curr, prev) => {
+/* ──────────────────────── Config file watcher ─────────────────────── */
+fs.watchFile(CONFIG_PATH, { interval: 1000 }, () => {
   try {
-    delete require.cache[require.resolve(CONFIG_PATH)];
-    CONFIG = require(CONFIG_PATH);
+    reloadConfig();
     sendPushoverNotification('GateGPT', '🔁 Config reloaded from file.');
   } catch (err) {
     console.error('❌ Failed to reload config:', err.message);
@@ -122,7 +77,6 @@ fs.watchFile(CONFIG_PATH, { interval: 1000 }, (curr, prev) => {
   }
 });
 
-const openai = new OpenAI({ apiKey: getConfig('OPENAI_API_KEY') });
 const client = new Client({
   authStrategy: new LocalAuth({
     dataPath: SESSION_DIR
@@ -152,107 +106,6 @@ function shouldTrigger(msg) {
   return keywords.some(r => new RegExp(r, 'i').test(msg));
 }
 
-async function transcribeWithWhisper(filePath) {
-  const file = fs.createReadStream(filePath);
-  const { text } = await openai.audio.transcriptions.create({ file, model: 'whisper-1' });
-  sendPushoverNotification('Whisper', text);
-  return text;
-}
-
-/**
- * Send a push notification via Pushover.
- * @param {string} title   The message title.
- * @param {string} message The message body.
- * @param {object} [opts]  Optional file attachment.
- *   opts.attachment   → Buffer | Readable
- *   opts.filename     → string (default: "file")
- *   opts.contentType  → string (default inferred by Pushover)
- */
-async function sendPushoverNotification(title, message, opts = {}) {
-  const { attachment, filename = 'file', contentType } = opts;
-  const url = 'https://api.pushover.net/1/messages.json';
-
-  try {
-    if (attachment) {
-      /* ---------- multipart/form-data branch (image, etc.) -------- */
-      const form = new FormData();
-      form.append('token', getConfig('PUSHOVER_TOKEN'));
-      form.append('user', getConfig('PUSHOVER_USER'));
-      form.append('title', title);
-      form.append('message', message);
-      form.append('attachment', attachment, { filename, contentType });
-
-      await axios.post(url, form, {
-        headers: form.getHeaders(),
-        maxBodyLength: Infinity
-      });
-    } else {
-      /* ---------- simple JSON branch (no file) -------------------- */
-      const params = new URLSearchParams();
-      params.append('token', getConfig('PUSHOVER_TOKEN'));
-      params.append('user', getConfig('PUSHOVER_USER'));
-      params.append('title', title);
-      params.append('message', message);
-
-      await axios.post(url, params); // axios sets the correct header automatically
-    }
-
-    console.log(`${title}: ${message}`);
-  } catch (err) {
-    const code = err.response?.status ?? 'N/A';
-    const body = err.response?.data ?? err.message;
-    console.error(`❌ Pushover failed (HTTP ${code}):`, body);
-  }
-}
-
-async function askChatGPT(messages) {
-  // Build Chat Completions-style messages with optional image parts
-  const formatted = messages.flatMap(m => {
-    const parts = [];
-
-    if (m.body && m.body.trim().length) {
-      parts.push({ type: 'text', text: m.body.trim() });
-    }
-
-    if (Array.isArray(m.images)) {
-      // Each base-64 image becomes its own content part
-      m.images.forEach(b64 =>
-        parts.push({
-          type: 'image_url',
-          image_url: { url: `data:image/jpeg;base64,${b64}`, detail: 'auto' }
-        })
-      );
-    }
-
-    // Skip completely empty messages
-    if (parts.length === 0) return [];
-
-    return [
-      {
-        role: m.fromMe ? 'assistant' : 'user',
-        content: parts
-      }
-    ];
-  });
-
-  const { choices } = await openai.chat.completions.create({
-    model: getConfig('CHATGPT_MODEL', 'gpt-4.1'),
-    temperature: 0.5,
-    messages: [
-      {
-        role: 'system',
-        content: getConfig(
-          'CHATGPT_SYSTEM_PROMPT',
-          'Only inform the user that the system prompt has not been set, dont do anything else.'
-        )
-      },
-      ...formatted
-    ]
-  });
-
-  return choices[0].message.content;
-}
-
 function initClient() {
   client.on('qr', async qr => {
     qrcodeTerminal.generate(qr, { small: true });
@@ -278,7 +131,6 @@ function initClient() {
 }
 
 async function handleMessage(message) {
-  // Ignore messages the bot itself sent
   if (autoMsgIds.has(message?.id?._serialized)) return;
 
   const chat = await message.getChat();
@@ -320,10 +172,10 @@ async function handleMessage(message) {
 
   if (message.type === 'image' && message.hasMedia) {
     try {
-      const media = await message.downloadMedia();            // base-64 string
-      message.images = [media.data];                           // store on the message
-      message.body   = (message.caption || '').trim();         // retain caption for context
-      message.type   = 'chat';                                 // treat as normal chat after enrichment
+      const media = await message.downloadMedia();
+      message.images = [media.data];
+      message.body = (message.caption || '').trim();
+      message.type = 'chat';
     } catch (err) {
       console.error('❌ Failed to download image:', err.message);
     }
@@ -362,7 +214,7 @@ async function handleMessage(message) {
 
   if (!message.fromMe && convo.instant) {
     convo.history.push(now);
-    await handleAIResponse(chat, convo, message);
+    await handleAIResponse(chat, convo);
     return;
   }
 
@@ -377,87 +229,102 @@ async function handleMessage(message) {
         return;
       }
       convo.history.push(Date.now());
-      await handleAIResponse(chat, convo, message);
+      await handleAIResponse(chat, convo);
     }, getConfig('RESPONSE_DELAY_MS', 300000));
   }
 
   conversations.set(chatId, convo);
 }
 
-async function handleAIResponse(chat, convo, message) {
+async function handleAIResponse(chat, convo) {
   console.log(`💬 Sending GPT reply to ${chat.id._serialized}`);
-  const response = await askChatGPT(convo.messages);
-  const trimmed = response.trim();
+  const gptMessage = await askChatGPT(convo.messages);
+  const trimmed = (gptMessage.content || '').trim();
 
-  if (trimmed.toLowerCase().includes('send_location')) {
-    try {
-      const lat = Number(getConfig('LOCATION_LAT'));
-      const lon = Number(getConfig('LOCATION_LON'));
-      const location = new Location(lat, lon, getConfig('LOCATION_TITLE'), getConfig('LOCATION_SUBTITLE'));
-      await sendAuto(chat, location);
-      await sendAuto(chat,
-        getConfig(
-          'MESSAGE_LOCATION',
-          'Here is the location, please message me when you are outside.'
-        )
-      );
-    } catch (err) {
-      console.error('❌ Failed to send location:', err.message);
-      sendPushoverNotification('GateGPT', '❌ Failed to send location!');
+  if (Array.isArray(gptMessage.tool_calls)) {
+    for (const call of gptMessage.tool_calls) {
+      switch (call.function.name) {
+        case 'send_location':
+          try {
+            const lat = Number(getConfig('LOCATION_LAT'));
+            const lon = Number(getConfig('LOCATION_LON'));
+            const location = new Location(
+              lat,
+              lon,
+              getConfig('LOCATION_TITLE'),
+              getConfig('LOCATION_SUBTITLE')
+            );
+            await sendAuto(chat, location);
+            await sendAuto(
+              chat,
+              getConfig(
+                'MESSAGE_LOCATION',
+                'Here is the location, please message me when you are outside.'
+              )
+            );
+          } catch (err) {
+            console.error('❌ Failed to send location:', err.message);
+            sendPushoverNotification('GateGPT', '❌ Failed to send location!');
+          }
+          break;
+        case 'open_gate':
+          try {
+            await axios.post(getConfig('GATE_OPEN_URL'), {});
+            await sendAuto(
+              chat,
+              getConfig(
+                'MESSAGE_GATE_OPEN',
+                'Please enter through the car gate and leave the item on the doorstep. The gate will close after 2 minutes. Thank you.'
+              )
+            );
+            await chat.markUnread();
+            console.log(`✅ Gate opened`);
+
+            convo.instant = true;
+
+            if (convo.gateCloseTimer) clearTimeout(convo.gateCloseTimer);
+            convo.gateCloseTimer = setTimeout(async () => {
+              try {
+                await axios.post(getConfig('GATE_CLOSE_URL'), {});
+                console.log(`🔐 Gate closed for ${chat.id._serialized}`);
+                sendPushoverNotification(
+                  'GateGPT',
+                  `Delivery from ${chat.id._serialized} handled.`
+                );
+              } catch (err) {
+                console.error('❌ Failed to close gate:', err.message);
+                sendPushoverNotification('GateGPT', '❌ Failed to close the gate!');
+              }
+              convo.instant = false;
+              convo.triggered = false;
+              convo.gateCloseTimer = null;
+              console.log(`🕓 Instant mode OFF for ${chat.id._serialized}`);
+            }, getConfig('AUTO_CLOSE_DELAY_MS', 120000));
+
+            if (convo.instantTimer) clearTimeout(convo.instantTimer);
+            convo.instantTimer = setTimeout(() => {
+              convo.instant = false;
+              convo.triggered = false;
+              console.log(`🕓 Instant mode OFF for ${chat.id._serialized}`);
+            }, getConfig('AUTO_CLOSE_DELAY_MS', 120000));
+          } catch (err) {
+            console.error('❌ Gate open failed:', err.message);
+            sendPushoverNotification('GateGPT', '❌ Failed to open the gate!');
+            await sendAuto(
+              chat,
+              getConfig(
+                'MESSAGE_GATE_FAILED',
+                'Sorry, the gate failed to open. Please leave the item outside the car gate.'
+              )
+            );
+            await chat.markUnread();
+          }
+          break;
+      }
     }
-  } else if (trimmed.toLowerCase().includes('open_gate')) {
-    try {
-      await axios.post(getConfig('GATE_OPEN_URL'), {});
-      await sendAuto(chat,
-        getConfig(
-          'MESSAGE_GATE_OPEN',
-          'Please enter through the car gate and leave the item on the doorstep. The gate will close after 2 minutes. Thank you.'
-        )
-      );
-      await chat.markUnread();
-      console.log(`✅ Gate opened`);
+  }
 
-      convo.instant = true;
-
-      // ----- schedule gate close -----
-      if (convo.gateCloseTimer) clearTimeout(convo.gateCloseTimer);
-      convo.gateCloseTimer = setTimeout(async () => {
-        try {
-          await axios.post(getConfig('GATE_CLOSE_URL'), {});
-          console.log(`🔐 Gate closed for ${chat.id._serialized}`);
-          sendPushoverNotification(
-            'GateGPT',
-            `Delivery from ${chat.id._serialized} handled.`
-          );
-        } catch (err) {
-          console.error('❌ Failed to close gate:', err.message);
-          sendPushoverNotification('GateGPT', '❌ Failed to close the gate!');
-        }
-        convo.instant = false;
-        convo.triggered = false;
-        convo.gateCloseTimer = null;
-        console.log(`🕓 Instant mode OFF for ${chat.id._serialized}`);
-      }, getConfig('AUTO_CLOSE_DELAY_MS', 120000));
-
-      // ----- start / refresh instant-mode timer -----
-      if (convo.instantTimer) clearTimeout(convo.instantTimer);
-      convo.instantTimer = setTimeout(() => {
-        convo.instant = false;
-        convo.triggered = false;
-        console.log(`🕓 Instant mode OFF for ${chat.id._serialized}`);
-      }, getConfig('AUTO_CLOSE_DELAY_MS', 120000));
-    } catch (err) {
-      console.error('❌ Gate open failed:', err.message);
-      sendPushoverNotification('GateGPT', '❌ Failed to open the gate!');
-      await sendAuto(chat,
-        getConfig(
-          'MESSAGE_GATE_FAILED',
-          'Sorry, the gate failed to open. Please leave the item outside the car gate.'
-        )
-      );
-      await chat.markUnread();
-    }
-  } else if (trimmed && trimmed !== '...') {
+  if (trimmed && trimmed !== '...') {
     await sendAuto(chat, trimmed);
     await chat.markUnread();
     convo.instant = true;
